Switch to OAuth backend with role mapping
This re-configures Grafana to use the OAuth backend (instead of
LDAP), which means we'll get automatic logins. Using a new feature of
Grafana, we can also make users admins if they're in ocfroot, and
"editors" if they are in ocfstaff.

This commit upgrades us to 6.5.0-beta1 to take advtange of role mapping.